requirements.txt
certifi==2021.10.8
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
numpy
olefile==0.46
Pillow==8.4.0
pyparsing==3.0.5
python-dateutil==2.8.2
six
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing-extensions
tensorboard
pytorch-fid
